rename: user페이지 디렉토리 생성
